requirements.txt
numpy~=1.23.4
python-dotenv~=0.21.0
pytest~=7.2.0
Pillow~=9.3.0
aiohttp~=3.8.3
aiofiles~=22.1.0
face-recognition~=1.3.0
clickhouse-driver==0.2.4
opencv-python==4.6.0.66
requests==2.28.1
yt-dlp==2022.11.11